Implemented adding and viewing of request by requester.
- My Requests tab now has a table that shows all the necessary details of each existing request made by the requester, only the requester's own requests will be shown.

Add New Request button now opens a modal where the requester will input details which are the supply's name and quantity per unit of that supply. Upon adding the supply to the list, a table below is shown of every supply recently added to the list. You can also delete an entry in the list. Upon submitting the request, a new request will now be added to the database along with the request_supplies junction table connecting multiple supplies to one request.